Fixes the vanishing datasets bug.  From:
http://groups.google.com/group/zfs-fuse/browse_thread/thread/b2feb8daafd5bb4b/eb31b8e9276d2b12

Thanks for the very long bug report ! ;-)

The fix this time is 3 characters long ! ;-)
Anyway, to give more details than just 3 characters : when listing the
available datasets, it allocates a structure at a given place, but forgets
to count the size of the path (don't ask me why !). So it doesn't allocate
enough, gets an "out of memory error" which you never see anywhere if you
don't debug it (we should probably change the code to make it more verbose
in case of error one day...).
So I just added 1024 to the default allocated size, and it seems to work
everywhere.

Attached the very short patch to fix the very big bug ! ;-)

- Rudd-O: I took the liberty to make the buffer 4K instead of 3K, for byte alignment and "to-be-on-the-safe-side" reasons.