Fixed groupmember (chown.t00) nonconformance to the NTFS-3G conformance tests.
Emmanuel: So the 1st patch, groupmember tries to reduce the number of calls to
groupmember since it's a slow function. I noticed zfs-fuse calls this
internally even when using root rights, which is a complete waste of time,
so I just added a few lines to return immediately if testing with root
access.
But since the cred_t was an opaque struct and I wanted to make things
easier, I also had to move the declaration of cred_t from cred_impl.h to
cred.h (read the patch, it's easier to understand when you read it).

Ref: http://groups.google.com/group/zfs-fuse/msg/c20b60d79b498671?hl=en