Full NTFS-3G conformance testing pass now.  Thanks to Emmanuel for this.
Emmanuel:

With this patch applied (and the one from this morning removed), every tests
pass.

To be sure we run the same thing, I had in conf :
fs="ZFS"
and in Makefile :
CFLAGS+=-DHAS_LCHMOD
# CFLAGS+=-DHAS_CHFLAGS
#CFLAGS+=-DHAS_LCHFLAGS
CFLAGS+=-DHAS_TRUNCATE64
# CFLAGS+=-DHAS_STAT64

And I got this at the end of the tests :
All tests successful.
Files=184, Tests=1957, 76 wallclock secs ( 9.31 cusr + 12.38 csys = 21.69
CPU)

Not very fast, but it works ! :)

REF:  http://groups.google.com/group/zfs-fuse/msg/63aacd137d535aa5?hl=en