Address the raidz1 problem in put_nvlist
when a drive is unavailable in a raidz1 pool and you try to import the pool,
you get an infinite loop on put_nvlist. Apparently this function was really
intended to fail in this case. So this quick fix tries to address the problem
without taking back the old bug of the disappearing pools. It's a lucky hack
for now, it seems to work everywhere but it should be improved one day...