mainly fixes for ztest to work with zfs-fuse
 - fixes almost all of the format warnings
 - make the "commit threshold exceeded" message a warning instead of a fatal
   since it can't be avoided if you have a single core cpu
 - Avoid to call list_destroy when the list wasn't even created (still a thread
   problem on single core cpu).

for txg.c:
sometimes it tries to destroy its callback list before the callbacks were
called.
It can probably happen on a multi core cpu too.
So this workaround just removes the callbacks if they are still in the list.

for spa_misc.c:
when compiled in debug mode (debug=2) it can try to display NULL blocks (bonus
blocks). The easiest solution is to avoid the crash in this case.