Disable a dangerous uiocopy in the middle of zfs_write
This uiocopy sends unrequested data to an ioctl socket, which can have
disastrous consequences (desync of sockets, buffers overwrites...).
It's a point which works clearly differently in opensolaris, but it seems
we can safety disable it for now. This fixes the bug where running a zfs
and a zpool command during io operations returned an error (which could
be a very dangerous error !!!).