Disable xattr support for now
xattr functions do not seem to be cached by fuse at all.
As a consequence, if you try :
dd if=/dev/zero of=zero bs=512 count=1000000
on a zfs filesystem
fuse will call getxattr for each of the 1000000 writes, which will
call lookup, which is extremely slow. With xattr support disabled
this dd is twice as fast here.
Should eventually be re-enabled later if a way to cache them is found.