Fix for io errors hanging zfs-fuse
There is a divergence here between opensolaris and zfs-fuse : in opensolaris in
case of EIO, zio is suspended and it waits for activity to be resumed. In
zfs-fuse nothing will ever send the signal to resume activity, so it can't
wait. This fix just prevents it from waiting for nothing.
Maybe we'll find something better later...