Ioctls transmit credentials
That's something opensolaris does.
Don't know if it will really be usefull, I guess a credential of 0 would be
enough in most cases. But transmiting the credentials here won't hurt, it's
just for the ioctls after all, not for all the fuse accesses. So added some
credential fields to zfsfuse_cmd_t and handled it correctly.