kstat support
This adds a /zfs-kstat directory with plenty of zfs attributes inside.
They are all read-only for now. Pass the argument "--no-kstat-mount" to
zfs-fuse if you don't want to see this directory. The directory is destroyed
when zfs-fuse exits.
Also added a modified version of arcstats.pl in contrib (modified to use the
zfs-kstat dir instead of the SUN::Solaris::Kstat module).