make unmounting "thread safe"
all this stuff is because fuse tries to be efficient and to unmount in the
background. The consequence is that if you do zfs umount -a && killall zfs-fuse
zfs-fuse is killed before all the fs have been unmounted (since now the ioctl
operations are threaded). At this point there is a race to unmount the
filesystems. These changes should make everything safe again.