powerpc support !
I received the main part from an anonymous mailer, for the defines in
isa_def.h.
I added a few parts from other places on the web (commented), and there is a
bug in spa_history_log, so I simply undefined _KERNEL in spa_history.c for
powerpc for now.
It's working quite well, tested on linux 2.6.18, powerpc 1.4 GHz (minimac).