failing vddevs and hot spares support
with fuse this support is much easier : now when a pool fails (or is
recovering), the command "zfs_pool_alert" is executed with the name of the pool
as argument. This commit includes a perl script for zfs_pool_alert which
handles hot spares if they are available.
Also this patch removes O_EXCL flag when opening a device, because this flag
prevents from using devices as hot spares (they are always opened and can be
shared between a few pools).