prevent calling autoconf/autoheader/automake from makefile
when browsing through git branches, the autoconf files could be touched by git
(even if not modified). The made make to call the autoconf tools but they fail
to regenerate the makefile most of the time, so it's better to stick with the
configure file taken from the tar.bz2 distribution.