cmd_listener ioctl threads now uses a queue and a thread pool
This prevents things from not scaling well (aborting under load).
The use of a proper queue allows for graceful shutdown (without races:
daemon hung in shutdown sleep(3)ing while >0.
If shutdown was in response to a signal, this leaves the process unable
to be terminated.
Simplified the handler model for a thread. A thread now needs nothing
besides the socket to talk to.
TODO: cleanup more and get rid of fds[MAX_CONNECTIONS] table?