SConstruct: incorporate suggested installation prefix as DESTDIR= and made it non-intrusive (tested)
> 717b16c Add support for DESTDIR
>
>     improved/modified and proposal in
>     1a530e6 SConstruct: incorporate suggested installation prefix as
>     prefix= and made it non-intrusive (tested
This is wrong. prefix and DESTDIR are two different things. Typically,
prefix, is defaulted to /usr/local and overridden by distros to /usr.
DESTDIR, on the other hand, tells the build system where to install the
stuff to, as a base directory instead of the root directory. They are
both separated because there can be variables in the source that need to
use the prefix (like, for example, when some programs load plugins from
$prefix/lib/my-app-plugins), in which case saying that the prefix is
$BUILD_DIR/some/directory is plain wrong.

Mike