bsd fix for hg commit 11816
From the mail :
Here is a patch adapted to the OpenSolaris tree.  We need it when
porting ZFS to FreeBSD but it seems to be useful for general consumption.

Currently:

 - vsnprintf() will not call va_end() and the va_list passed is not defined;
 - subsequent vsprintf() references the passed va_list again.

On FreeBSD/amd64, the latter vsprintf() would write undefined result,
which tends to cause random panics.

It turns out that the undefined behavior is defined in standard, so we
have added a pair of va_copy() and va_end() calls and operate on the
copy for the first vsnprintf().

This is needed only if 11816 is applied.