use O_SYNC with O_DIRECT
This was a silly problem, really.
The man page for open(2) clearly tells to use O_SYNC with O_DIRECT.
For an unknown reason zfs-fuse didn't do so, and it created a long standing bug
with ztest telling us "Commit callback threshold exceeded". Well simply using
this O_SYNC flag seems to fix the problem, ztest remains silent again (and so
the message becomes a fatal again).