add basic buffers for writes
I thought this was for dumb programs only, but apparently a lot of programs are
doing small writes which can be optimized by this.
The drawback is that the buffers are never flushed on timeout, they are flushed
only when the files are closed, written at a different offset, read, or if the
buffer grows over a size of 128k.