Readd something for getattr after all
The getattr behavior is really a fuse bug :
if a program tries to get the size of a file with stat before the file
is closed then fuse sends no message to flush the cache, and getattr
does not allow to retrieve the fi parameter passed to open.

So this is a new and better version of the add_info/free_info/get_info
functions. This time it affects only getattr, and does not flush the
cache in this case, it just tries to fix the size returned by stat based
on the offset of the current buffer and its size. Normally there is no
need to change zfsfuse_stat because of the flush callback now.
Also this version tries to keep the list of info as small as possible,
and use it as little as possible (there is a lock, but it won't be used
often !).

Tested with :
git commit -a in any git directory (failed with the version of
yesterday)
sudo apt-src install util-linux
git clone something

Also remove debug messages forgotten in the previous commit.