adjust pools uuid to util-linux patch
the format is now decimal and not hex, and we should try 1st to open by uuid
and then only by the pool name. This to be sure that it will work even if the
partition is still on the same device, but not at the same number (which can
happen after deleting a partition for example).

Notice also that you need an adjuted
/lib/udev/rules.d/60-persistent-storage.rules
which allows raid devices in /dev/by-uuid, and which display the UUID_SUB
record in this case because the pool UUID itself would not be very useful here,
we want the uuid of this partition or node.