an "improvement" for the buffers
If a file is opened for writing and uses buffers, and then opened for reading,
the 2nd open won't see the buffers so it won't be able to flush them. It can be
reproduced with a simple test program, 2 open, fread / fwrite.
I hesitate to call this a "bug" because so far I was unable to reproduce this
in a real situation, only my test program shows there could be a problem. So I
left a syslog message to see if this fix will be useful or not.